renamed component-services to component-service-s3
