+++ packages/component-dashboard/server/resolvers.js
@@ -1,5 +1,5 @@
 const logger = require('@pubsweet/logger')
-const { S3Storage } = require('@elifesciences/component-services')
+const { S3Storage } = require('@elifesciences/component-service-s3')
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
 const File = require('@elifesciences/component-model-file').model
