+++ packages/component-service-s3/index.js
