+++ packages/component-service-s3/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@elifesciences/component-services",
+  "name": "@elifesciences/component-service-s3",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT"
