+++ packages/component-service-s3/server/pubsub.js
