+++ packages/component-submission/server/resolvers/submitManuscript.js
@@ -5,7 +5,7 @@ const logger = require('@pubsweet/logger')
 const { mecaExport } = require('@elifesciences/component-meca')
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
-const { S3Storage } = require('@elifesciences/component-services')
+const { S3Storage } = require('@elifesciences/component-service-s3')
 const manuscriptInputSchema = require('../helpers/manuscriptInputValidationSchema')
 
 const { Notification } = require('../services')
